refactor: move screenshot filename helper to core
