+++ crates/axon-adapters/src/web_engine/screenshot.rs
@@ -383,45 +383,3 @@ where
             .unwrap_or(serde_json::Value::Null));
     }
 }
-
-/// Sanitize a URL into a safe screenshot filename.
-///
-/// Strips the scheme, replaces non-alphanumeric chars with hyphens,
-/// collapses runs of hyphens, trims edges, and truncates to 120 chars.
-pub fn url_to_screenshot_filename(url: &str, idx: usize) -> String {
-    let stripped = url
-        .strip_prefix("https://")
-        .or_else(|| url.strip_prefix("http://"))
-        .unwrap_or(url);
-
-    let sanitized: String = stripped
-        .chars()
-        .map(|c| if c.is_ascii_alphanumeric() { c } else { '-' })
-        .collect();
-
-    // Collapse consecutive hyphens and trim leading/trailing hyphens.
-    let mut collapsed = String::with_capacity(sanitized.len());
-    let mut prev_hyphen = true; // Start true to trim leading hyphens.
-    for c in sanitized.chars() {
-        if c == '-' {
-            if !prev_hyphen {
-                collapsed.push('-');
-            }
-            prev_hyphen = true;
-        } else {
-            collapsed.push(c);
-            prev_hyphen = false;
-        }
-    }
-    let collapsed = collapsed.trim_end_matches('-');
-
-    // Truncate to a reasonable filename length.
-    let max_name = 120;
-    let name = if collapsed.len() > max_name {
-        &collapsed[..max_name]
-    } else {
-        collapsed
-    };
-
-    format!("{idx:04}-{name}.png")
-}
