+++ crates/axon-cli/Cargo.toml
@@ -9,7 +9,6 @@ license = "MIT"
 [dependencies]
 axon-api = { path = "../axon-api" }
 axon-core = { path = "../axon-core" }
-axon-adapters = { path = "../axon-adapters" }
 axon-jobs = { path = "../axon-jobs" }
 axon-services = { path = "../axon-services" }
 axon-mcp = { path = "../axon-mcp" }
