+++ crates/axon-cli/src/commands/screenshot/util.rs
@@ -1,11 +1,6 @@
 use axon_core::config::Config;
 use std::error::Error;
 
-// Re-export for tests only — canonical implementation lives in src/crawl/screenshot.rs.
-// Services/MCP import from there directly; only test modules in this subtree use the re-export.
-#[cfg(test)]
-pub(crate) use axon_adapters::web_engine::screenshot::url_to_screenshot_filename;
-
 /// Validate that Chrome is configured before attempting a screenshot.
 pub(super) fn require_chrome(cfg: &Config) -> Result<(), Box<dyn Error>> {
     if cfg.chrome_remote_url.is_none() {
