+++ crates/axon-cli/src/commands/screenshot/util_tests.rs
@@ -1,5 +1,6 @@
 use super::*;
 use axon_core::config::Config;
+use axon_core::paths::url_to_screenshot_filename;
 
 // --- url_to_screenshot_filename ---
