+++ docs/reference/crate-dependency-graph.md
@@ -5,7 +5,7 @@ Generated by `cargo xtask gen-dep-graph`. Normal + build dependencies over works
 - `axon-adapters` → `axon-api`, `axon-core`, `axon-error`, `axon-extract`, `axon-observe`, `axon-parse`
 - `axon-api` → `axon-error`
 - `axon-authz` → `axon-api`, `axon-error`
-- `axon-cli` → `axon-adapters`, `axon-api`, `axon-core`, `axon-jobs`, `axon-mcp`, `axon-services`, `axon-web`
+- `axon-cli` → `axon-api`, `axon-core`, `axon-jobs`, `axon-mcp`, `axon-services`, `axon-web`
 - `axon-core` → `axon-api`
 - `axon-document` → `axon-api`, `axon-core`, `axon-parse`
 - `axon-embedding` → `axon-api`, `axon-error`, `axon-observe`
