+++ docs/reference/public-api-surface.md
@@ -274,7 +274,6 @@ Generated by `cargo xtask gen-public-api`. Lists each workspace crate's **crate-
 - `web_engine::scrape::select_output` (fn)
 - `web_engine::screenshot` (mod)
 - `web_engine::screenshot::spider_screenshot_with_options` (fn)
-- `web_engine::screenshot::url_to_screenshot_filename` (fn)
 - `youtube` (mod)
 - `youtube::MODULE_NAME` (const)
 - `youtube::YoutubeSourceAdapter` (struct)
@@ -1350,6 +1349,7 @@ Generated by `cargo xtask gen-public-api`. Lists each workspace crate's **crate-
 - `paths::ensure_private_dir_async` (fn)
 - `paths::open_private_append` (fn)
 - `paths::path_basename` (fn)
+- `paths::url_to_screenshot_filename` (fn)
 - `redact` (mod)
 - `redact::BARE_SECRET_TOKEN_PREFIXES` (use)
 - `redact::DefaultRedactor` (use)
